Update aws_accounts gql and typed_query to fit terraform_tgw_attachments
