--- reconcile/gql_definitions/aws_accounts/aws_accounts.gql
@@ -1,76 +1,0 @@
-# qenerate: plugin=pydantic_v1
-
-query AWSAccounts(
-  $name: String,
-  $uid: String,
-  $ecrs: Boolean!,
-  $reset_passwords: Boolean!,
-  $sharing: Boolean!,
-  $terraform_state: Boolean!,
-) {
-  accounts: awsaccounts_v1
-  (
-    name: $name
-    uid: $uid
-  )
-  {
-    path
-    name
-    uid
-    terraformUsername
-    consoleUrl
-    resourcesDefaultRegion
-    supportedDeploymentRegions
-    providerVersion
-    accountOwners {
-      name
-      email
-    }
-    automationToken {
-      ... VaultSecret
-    }
-    garbageCollection
-    enableDeletion
-    deletionApprovals {
-      type
-      name
-      expiration
-    }
-    disable {
-      integrations
-    }
-    deleteKeys
-    resetPasswords @include(if: $reset_passwords) {
-      user {
-        org_username
-      }
-      requestId
-    }
-    premiumSupport
-    ecrs @include(if: $ecrs) {
-      region
-    }
-    partition
-    sharing @include(if: $sharing) {
-      provider
-      account {
-        name
-        uid
-        supportedDeploymentRegions
-      }
-      ... on AWSAccountSharingOptionAMI_v1 {
-        regex
-        region
-      }
-    }
-    terraformState @include(if: $terraform_state) {
-      provider
-      bucket
-      region
-      integrations {
-        key
-        integration
-      }
-    }
-  }
-}
