+++ reconcile/gql_definitions/fragments/aws_account_common.py
@@ -1,0 +1,65 @@
+"""
+Generated by qenerate plugin=pydantic_v1. DO NOT MODIFY MANUALLY!
+"""
+from collections.abc import Callable  # noqa: F401 # pylint: disable=W0611
+from datetime import datetime  # noqa: F401 # pylint: disable=W0611
+from enum import Enum  # noqa: F401 # pylint: disable=W0611
+from typing import (  # noqa: F401 # pylint: disable=W0611
+    Any,
+    Optional,
+    Union,
+)
+
+from pydantic import (  # noqa: F401 # pylint: disable=W0611
+    BaseModel,
+    Extra,
+    Field,
+    Json,
+)
+
+from reconcile.gql_definitions.fragments.vault_secret import VaultSecret
+
+
+class ConfiguredBaseModel(BaseModel):
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class OwnerV1(ConfiguredBaseModel):
+    name: str = Field(..., alias="name")
+    email: str = Field(..., alias="email")
+
+
+class DeletionApprovalV1(ConfiguredBaseModel):
+    q_type: str = Field(..., alias="type")
+    name: str = Field(..., alias="name")
+    expiration: str = Field(..., alias="expiration")
+
+
+class DisableClusterAutomationsV1(ConfiguredBaseModel):
+    integrations: Optional[list[str]] = Field(..., alias="integrations")
+
+
+class AWSAccountCommon(ConfiguredBaseModel):
+    path: str = Field(..., alias="path")
+    name: str = Field(..., alias="name")
+    uid: str = Field(..., alias="uid")
+    terraform_username: Optional[str] = Field(..., alias="terraformUsername")
+    console_url: str = Field(..., alias="consoleUrl")
+    resources_default_region: str = Field(..., alias="resourcesDefaultRegion")
+    supported_deployment_regions: Optional[list[str]] = Field(
+        ..., alias="supportedDeploymentRegions"
+    )
+    provider_version: str = Field(..., alias="providerVersion")
+    account_owners: list[OwnerV1] = Field(..., alias="accountOwners")
+    automation_token: VaultSecret = Field(..., alias="automationToken")
+    garbage_collection: Optional[bool] = Field(..., alias="garbageCollection")
+    enable_deletion: Optional[bool] = Field(..., alias="enableDeletion")
+    deletion_approvals: Optional[list[DeletionApprovalV1]] = Field(
+        ..., alias="deletionApprovals"
+    )
+    disable: Optional[DisableClusterAutomationsV1] = Field(..., alias="disable")
+    delete_keys: Optional[list[str]] = Field(..., alias="deleteKeys")
+    premium_support: bool = Field(..., alias="premiumSupport")
+    partition: Optional[str] = Field(..., alias="partition")
