+++ reconcile/test/test_typed_queries/terraform_tgw_attachments/test_get_aws_accounts.py
@@ -3,11 +3,13 @@ from unittest.mock import ANY
 
 import pytest
 
-from reconcile.gql_definitions.aws_accounts.aws_accounts import (
+from reconcile.gql_definitions.terraform_tgw_attachments.aws_accounts import (
     AWSAccountsQueryData,
     AWSAccountV1,
 )
-from reconcile.typed_queries.aws_accounts import get_aws_accounts
+from reconcile.typed_queries.terraform_tgw_attachments.aws_accounts import (
+    get_aws_accounts,
+)
 from reconcile.utils.gql import GqlApi
 
 
@@ -65,11 +67,6 @@ def test_get_aws_accounts_with_default_variables(
     assert result == aws_accounts.accounts
     expected_variables = {
         "name": None,
-        "uid": None,
-        "ecrs": True,
-        "reset_passwords": False,
-        "sharing": False,
-        "terraform_state": False,
     }
     gql_api.query.assert_called_once_with(ANY, variables=expected_variables)
 
@@ -85,35 +82,20 @@ def test_get_aws_accounts_when_no_data(
     assert result == []
 
 
-def test_get_aws_accounts_with_variables(
+def test_get_aws_accounts_with_name(
     gql_api_builder: Callable[..., GqlApi],
     aws_accounts: AWSAccountsQueryData,
 ) -> None:
     gql_api = gql_api_builder(aws_accounts.dict(by_alias=True))
     name = "some-name"
-    uid = "some-uid"
-    ecrs = False
-    reset_passwords = True
-    sharing = True
-    terraform_state = True
 
     result = get_aws_accounts(
         gql_api,
         name=name,
-        uid=uid,
-        ecrs=ecrs,
-        reset_passwords=reset_passwords,
-        sharing=sharing,
-        terraform_state=terraform_state,
     )
 
     assert result == aws_accounts.accounts
     expected_variables = {
         "name": name,
-        "uid": uid,
-        "ecrs": ecrs,
-        "reset_passwords": reset_passwords,
-        "sharing": sharing,
-        "terraform_state": terraform_state,
     }
     gql_api.query.assert_called_once_with(ANY, variables=expected_variables)
